Refactor logic into storage and metrics modules
